fix tests on interface change

+++ src/test/kotlin/com/hagoapp/f2t/excel/TypedExcelReadTest.kt
@@ -59,7 +59,7 @@ class TypedExcelReadTest {
             parser.determiner = FileTypeDeterminer(determiner)
             parser.addObserver(observer)
             parser.parse()
-            Assertions.assertEquals(testConfig.expect.rowCount, observer.rowCount)
+            Assertions.assertEquals(testConfig.expect.rowCount.toLong(), observer.rowCount)
             Assertions.assertEquals(testConfig.expect.columnCount, observer.columns.size)
             Assertions.assertEquals(
                 testConfig.expect.types,
